feat: add runCli entrypoint helper to drop consumer boilerplate
Shipping a `runCli(arguments, {fileRendererBuilder})` helper in
`package:space_gen/space_gen.dart` so a custom consumer entrypoint
is one line. The helper parses the standard `--in`/`--out`/
`--verbose`/`--openapi` flags, resolves the bundled templates
directory, wires up a Logger, and calls loadAndRenderSpec.

`bin/space_gen.dart` collapses to:

    Future<int> main(List<String> args) => runCli(args);

A consumer swapping just the FileRenderer is:

    Future<int> main(List<String> args) =>
        runCli(args, fileRendererBuilder: MyRenderer.new);

Consumers that need new flags or non-CLI argument sources can skip
runCli and build a GeneratorConfig themselves.
