+++ lib/space_gen.dart
@@ -1,3 +1,4 @@
+export 'src/cli.dart' show runCli;
 export 'src/logger.dart' show Logger, logger, runWithLogger, setVerboseLogging;
 export 'src/quirks.dart' show Quirks;
 export 'src/render.dart'
